--- .github/workflows Release.yml
@@ -1,59 +1,0 @@
-name: Release Chirp
-
-on:
-  push:
-    branch:
-      - main   # Triggers when you push or merges to main
-
-jobs:
-  build:
-    runs-on: ubuntu-latest
-
-    steps:
-      - name: Checkout repository
-        uses: actions/checkout@v4
-
-      - name: Setup .NET
-        uses: actions/setup-dotnet@v4
-        with:
-          dotnet-version: '9.0.x'   # Change if you use a different version
-
-      - name: Restore dependencies
-        run: dotnet restore
-
-      - name: Publish single-file executable
-        run: |
-          dotnet publish src/Chirp.Web/Chirp.Web.csproj \
-            -c Release \
-            -r linux-x64 \
-            --self-contained true \
-            /p:PublishSingleFile=true \
-            /p:IncludeAllContentForSelfExtract=true \
-            -o ./publish
-
-      - name: Upload artifact
-        uses: actions/upload-artifact@v4
-        with:
-          name: chirp-executable
-          path: ./publish/*
-
-  release:
-    runs-on: ubuntu-latest
-    needs: build
-
-    steps:
-      - name: Download build artifact
-        uses: actions/download-artifact@v4
-        with:
-          name: chirp-executable
-          path: ./publish
-
-      - name: Create GitHub Release
-        uses: softprops/action-gh-release@v2
-        with:
-          tag_name: v${{ github.run_number }} # automatic version number
-          name: "Release v${{ github.run_number }}" # release title
-          body: "Automatic release generated by main merge"
-          files: ./publish/*
-        env:
-          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
