--- test/Chirp.Razor.Tests/UnitTest1.cs
@@ -1,22 +1,0 @@
-﻿using Microsoft.Data.Sqlite;
-using Xunit;
-
-namespace Chirp.Razor.Tests;
-using Chirp.Razor.wwwroot;
-
-
-public class UnitTest1
-{
-    [Fact]
-    public void Read_ReturnsCheeps_FromInMemoryDatabase()
-    {
-        // Arrange
-        
-
-        // Act
-        
-
-        // Assert
-        
-    }
-}
