+++ test/Chirp.Repositories.Tests/Chirp.Repositories.Tests.sln
@@ -1,0 +1,24 @@
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.5.2.0
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Chirp.Repositories.Tests", "Chirp.Repositories.Tests.csproj", "{40EF6A0C-C61B-6A86-4463-F692C9EDF9D7}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{40EF6A0C-C61B-6A86-4463-F692C9EDF9D7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{40EF6A0C-C61B-6A86-4463-F692C9EDF9D7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{40EF6A0C-C61B-6A86-4463-F692C9EDF9D7}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{40EF6A0C-C61B-6A86-4463-F692C9EDF9D7}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {1DEE3A78-8C4C-4999-A8A1-98DE7B59264E}
+	EndGlobalSection
+EndGlobal
