+++ test/ChirpEndToEndTests/ChirpEndToEndTests.sln
@@ -2,7 +2,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.5.2.0
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Chirp.DB.Tests", "Chirp.DB.Tests.csproj", "{254B1232-8EE2-12B6-F975-C44E992F92AA}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ChirpEndToEndTests", "ChirpEndToEndTests.csproj", "{47D64BA8-B774-1DAA-326C-283FA55E7889}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -10,15 +10,15 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{254B1232-8EE2-12B6-F975-C44E992F92AA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{254B1232-8EE2-12B6-F975-C44E992F92AA}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{254B1232-8EE2-12B6-F975-C44E992F92AA}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{254B1232-8EE2-12B6-F975-C44E992F92AA}.Release|Any CPU.Build.0 = Release|Any CPU
+		{47D64BA8-B774-1DAA-326C-283FA55E7889}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{47D64BA8-B774-1DAA-326C-283FA55E7889}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{47D64BA8-B774-1DAA-326C-283FA55E7889}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{47D64BA8-B774-1DAA-326C-283FA55E7889}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {8D14A301-D50E-488E-8ABA-37F2BA100A38}
+		SolutionGuid = {70FEFF01-5440-4698-A367-997005DE8D32}
 	EndGlobalSection
 EndGlobal
