--- test/test.sln
@@ -1,30 +1,0 @@
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.5.2.0
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Chirp.DB.Tests", "Chirp.DB.Tests\Chirp.DB.Tests.csproj", "{8F6A35A8-8E8B-383B-1D09-A9B01CF6FFB9}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Chirp.Razor.Tests", "Chirp.Razor.Tests\Chirp.Razor.Tests.csproj", "{836BEC8A-EEFA-EDAF-18D9-99591C7F3378}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{8F6A35A8-8E8B-383B-1D09-A9B01CF6FFB9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{8F6A35A8-8E8B-383B-1D09-A9B01CF6FFB9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8F6A35A8-8E8B-383B-1D09-A9B01CF6FFB9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{8F6A35A8-8E8B-383B-1D09-A9B01CF6FFB9}.Release|Any CPU.Build.0 = Release|Any CPU
-		{836BEC8A-EEFA-EDAF-18D9-99591C7F3378}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{836BEC8A-EEFA-EDAF-18D9-99591C7F3378}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{836BEC8A-EEFA-EDAF-18D9-99591C7F3378}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{836BEC8A-EEFA-EDAF-18D9-99591C7F3378}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {C4BBD3A2-7101-4F3C-ABE3-1897ACAC4701}
-	EndGlobalSection
-EndGlobal
